Restructure project and add Meson build system support
