+++ .gitignore
@@ -40,6 +40,8 @@
 /docs/_xml
 
 # Generated files
-/src/include/rawrtcc_configuration.h
 /docs/Doxyfile
 /docs/conf.py
+
+# Subprojects
+/subprojects/re
