+++ include/meson.build
@@ -1,0 +1,3 @@
+# Install headers
+install_headers(files('rawrtcc.h'))
+subdir('rawrtcc')
