+++ src/certificate/meson.build
@@ -1,0 +1,1 @@
+sources += files('certificate.c')
