+++ src/meson.build
@@ -1,0 +1,5 @@
+sources = []
+
+subdir('certificate')
+subdir('message_buffer')
+subdir('utils')
