+++ src/message_buffer/meson.build
@@ -1,0 +1,1 @@
+sources += files('buffer.c')
