+++ src/utils/meson.build
@@ -1,0 +1,1 @@
+sources += files('utils.c')
